plan: let a half credit be a half credit
`/plan` returned a 500 for any student who entered a 1.5-credit course, and it is
not an edge case: MSEM, TCTM and TCHS are built out of them — a 16.5-credit core
plus a 1.5-credit internship or capstone — and MSEM's internship is a single
1.5-credit *required* course. Three of the twenty-two encoded degrees were
unusable from the first thing the product asks a student to do. The blast radius
is wider than those three, because `loader.load_catalog_courses` hands every
degree all 749 catalog courses, so all 129 fractional ones are in scope for
everyone.

The cause was an annotation that had been wrong for a while and could not show
it: `CourseRule.credits` was already float, `PlanResult` and `PlanOut` declared
`int`, and nothing between them checked. Pydantic did, two layers away, by
refusing to coerce 1.5. `/plan` is what the planner, the rail and the chat card
all read, so the failure took out three surfaces at once.

The float goes all the way through — PlanResult, PlanOut, RequirementProgress,
ReadinessResponse, CourseNeed, ProfileEntry, and readiness' accumulators. Only
`percent_complete` stays int, because it is a rounded percentage and not a
credit count.

Which exposed the second defect, this one already live. `Requirement.min_credits`
has always been Float, so every `credits` requirement in all 22 degrees rendered

    Electives: 15.0 credit(s) short
    0 of 15.0 credits so far. Select 15 credits across the ...

with the tool's own number disagreeing with the bulletin's quoted one inside a
single sentence, on the surface whose whole claim is that figures have
provenance. `planning/format.fmt_credits` renders 15.0 as "15" and leaves 1.5
alone. It deliberately does not round: 0.25 stays 0.25, because a total the
student cannot reconcile against the bulletin is worth more than a tidy number
that quietly lost a quarter credit.

And a third, in the browser. `useCountUp` rounded unconditionally, so a 1.5
credit total animated up and landed on 2 — wrong in the direction that tells a
student they are further along than they are. It now steps in whole numbers for
a whole-number target, exactly as before, and in tenths otherwise.

Verified against the dev database, not only in fixtures: MSEM1-GC 1100 as
completed went from ValidationError to credits_completed=1.5, the same for the
what-if path, and the finding text on three real degrees lost its trailing zero.
441 passed.

+++ api/app/planning/types.py
@@ -100,10 +100,17 @@ class PlanResult:
     """Everything the engine concluded, ready for the model to narrate."""
 
     findings: list[Finding] = field(default_factory=list)
-    credits_completed: int = 0
-    credits_in_progress: int = 0
-    credits_planned: int = 0
-    credits_required: int = 0
+    # Float, because half credits are real and they are not decoration: three encoded
+    # degrees (MSEM, TCTM, TCHS) are built out of 1.5-credit courses, and MSEM's internship
+    # — a single 1.5-credit *required* course — means the first thing such a student enters
+    # produces a fractional total. These were annotated `int` while `CourseRule.credits` was
+    # already float, so the lie was invisible here and surfaced two layers away as a 500
+    # from `/plan` when pydantic refused to coerce 1.5. `credits_required` stays float for
+    # symmetry with the other three even though its column is Integer today.
+    credits_completed: float = 0.0
+    credits_in_progress: float = 0.0
+    credits_planned: float = 0.0
+    credits_required: float = 0.0
 
     def add(self, finding: Finding) -> None:
         self.findings.append(finding)
